Add camera drag controller
