+++ CodexTest/Assets/Scripts/Presentation/CameraController.cs.meta
@@ -1,11 +1,11 @@
 fileFormatVersion: 2
-guid: 2726c9907b3bf6f4c9b4162328ac503f
+guid: 767007e9dd7b4d62a18c9e30751e105d
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
   defaultReferences: []
   executionOrder: 0
   icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant: 
+  userData:
+  assetBundleName:
+  assetBundleVariant:
